requirements.txt
kafka-python === 2.0.2
pytest === 6.2.2
aiven-client === 2.10.0
psycopg2-binary === 2.8.6
pylint
flake8
requests
